conflict rez
